+++ package.json
@@ -60,11 +60,11 @@
     "eslint-config-next": "13.4.6",
     "eslint-config-prettier": "^8.8.0",
     "eslint-import-resolver-typescript": "^3.5.5",
-    "eslint-plugin-import": "2.27.5",
-    "eslint-plugin-jest": "^27.2.2",
-    "eslint-plugin-prettier": "5.0.0",
-    "eslint-plugin-simple-import-sort": "^10.0.0",
     "eslint-plugin-storybook": "^0.6.12",
+    "eslint-plugin-jest": "^27.2.2",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-simple-import-sort": "^10.0.0",
+    "eslint-plugin-import": "2.27.5",
     "eslint-plugin-unused-imports": "3.0.0",
     "given2": "^2.1.7",
     "husky": "^8.0.0",
@@ -72,9 +72,8 @@
     "jest-plugin-context": "^2.9.0",
     "lint-staged": "^13.2.3",
     "postcss": "^8.4.18",
-    "prettier": "3.0.0",
+    "prettier": "^2.8.8",
     "storybook": "^7.0.26",
     "tailwindcss": "^3.2.1"
   }
 }
-
